fix: 상세페이지 금주의 cs, 이전 cs 분리
