+++ src/components/cs/SubmittedDetailBox.tsx
@@ -3,6 +3,8 @@
 import AnswerCard from './AnswerCard'
 import { useSubmittedDetailBox } from '@/hooks/cs/useSubmittedDetailBox'
 import { useInfiniteScroll } from '@/hooks/cs/useInfiniteScroll'
+import { useTodayCsQuery } from '@/api/cs/queries'
+import { useProblemDate } from '@/hooks/cs/useProblemDate'
 
 interface SubmittedDetailBoxProps {
   id: string
@@ -21,16 +23,29 @@ export default function SubmittedDetailBox({ id }: SubmittedDetailBoxProps) {
     refreshAnswers, // 새로고침 함수 추가
   } = useSubmittedDetailBox({ id })
 
+  // 오늘의 CS 문제 조회 (금주의 CS 판단용)
+  const { data: todayCs } = useTodayCsQuery()
+
   const { ref } = useInfiniteScroll({
     fetchNextPage,
     hasNextPage,
     isFetchingNextPage,
   })
 
+  // 현재 문제가 금주의 CS인지 판단
+  const isCurrentWeekCS = todayCs?.problemId === Number(id)
+
+  // 공통 훅 사용
+  const { getProblemDate } = useProblemDate(problemDetail?.updatedAt || '')
+
   if (problemLoading) {
     return (
       <div className="max-w-[1200px] mx-auto mt-[3.56rem]">
-        <h1 className="text-[2rem] font-bold mb-5"> CS 문제</h1>
+        <div className="flex items-center space-x-4 mb-5">
+          <h1 className="text-[2rem] font-bold">
+            {isCurrentWeekCS ? '금주의 CS' : '이전 CS'}
+          </h1>
+        </div>
         <div className="border border-gray bg-filterbg px-6 py-8 rounded-xl text-xl font-semibold mb-10">
           <p>문제를 불러오는 중...</p>
         </div>
@@ -41,7 +56,11 @@ export default function SubmittedDetailBox({ id }: SubmittedDetailBoxProps) {
   if (problemError || !problemDetail) {
     return (
       <div className="max-w-[1200px] mx-auto mt-[3.56rem]">
-        <h1 className="text-[2rem] font-bold mb-5">CS 문제</h1>
+        <div className="flex items-center space-x-4 mb-5">
+          <h1 className="text-[2rem] font-bold">
+            {isCurrentWeekCS ? '금주의 CS' : '이전 CS'}
+          </h1>
+        </div>
         <div className="border border-red-300 bg-red-50 px-6 py-8 rounded-xl text-xl font-semibold mb-10">
           <p>문제를 불러오는데 실패했습니다.</p>
         </div>
@@ -51,8 +70,12 @@ export default function SubmittedDetailBox({ id }: SubmittedDetailBoxProps) {
 
   return (
     <div className="max-w-[1200px] mx-auto mt-[3.56rem]">
-      <h1 className="text-[2rem] font-bold mb-5">CS 문제</h1>
-      {/* <p className="text-primary text-lg mb-3">2025년 03월 28일</p> */}
+      <div className="flex items-center space-x-4 mb-5">
+        <h1 className="text-[2rem] font-bold">
+          {isCurrentWeekCS ? '금주의 CS' : '이전 CS'}
+        </h1>
+        <p className="text-lg text-gray">{getProblemDate()}</p>
+      </div>
 
       <div className="border border-gray bg-filterbg px-6 py-8 rounded-xl text-xl font-semibold mb-10">
         <p>{problemDetail.content}</p>
